+++ hes/src/test/java/com/memmcol/hes/domain/network/NetworkWriteServiceTest.java
@@ -139,8 +139,8 @@ public class NetworkWriteServiceTest {
         assertEquals("success", result.get("status"));
         assertEquals(ipPorts, result.get("ipPorts"));
 
-        // Verify Port written first (Class 41, Attr 2)
-        verify(dlmsReaderUtils).writeAttribute(eq(dlmsClient), eq(serial), eq("0.11.25.0.0.255"), eq(41), eq(2), eq(port), eq(DataType.UINT16));
+        // Verify Port written first (Class 41, Attr 2) as UINT32
+        verify(dlmsReaderUtils).writeAttribute(eq(dlmsClient), eq(serial), eq("0.11.25.0.0.255"), eq(41), eq(2), eq((long)port), eq(DataType.UINT32));
         // Verify IP written second (Class 45, Attr 5)
         verify(dlmsReaderUtils).writeAttribute(eq(dlmsClient), eq(serial), eq("0.11.25.4.0.255"), eq(45), eq(5), any(byte[].class), eq(DataType.OCTET_STRING));
     }
